fix mutilSig sort
