+++ ontolib/src/main/java/com/github/ontio/common/Address.java
@@ -20,6 +20,7 @@
 package com.github.ontio.common;
 
 import com.alibaba.fastjson.JSON;
+import com.github.ontio.core.program.Program;
 import com.github.ontio.core.scripts.ScriptBuilder;
 import com.github.ontio.core.scripts.ScriptOp;
 import com.github.ontio.crypto.KeyType;
@@ -102,49 +103,9 @@ public class Address extends UIntBase implements Comparable<Address> {
     }
 
     public static Address addressFromMultiPubKeys(int m, byte[]... publicKeys) throws Exception {
-        if (m <= 0 || m > publicKeys.length || publicKeys.length > 24) {
-            throw new SDKException(ErrorCode.ParamError);
-        }
-        try (ScriptBuilder sb = new ScriptBuilder()) {
-            sb.push(BigInteger.valueOf(m));
-            Arrays.sort(publicKeys, new Comparator<byte[]>() {
-                @Override
-                public int compare(byte[] o1, byte[] o2) {
-                    if (KeyType.fromPubkey(o1).getLabel() != KeyType.fromPubkey(o2).getLabel()) {
-                        return KeyType.fromPubkey(o1).getLabel() > KeyType.fromPubkey(o2).getLabel() ? 1 : -1;
-                    }
-                    switch (KeyType.fromPubkey(o1)) {
-                        case SM2:
-                            byte[] p = new byte[33];
-                            System.arraycopy(o1, 2, p, 0, p.length);
-                            o1 = p;
-                            byte[] p2 = new byte[33];
-                            System.arraycopy(o2, 2, p2, 0, p2.length);
-                            o2 = p2;
-                            ECPoint smPk1 = ECC.sm2p256v1.getCurve().decodePoint(o1);
-                            ECPoint smPk2 = ECC.sm2p256v1.getCurve().decodePoint(o2);
-                            return ECC.compare(smPk1, smPk2);
-                        case ECDSA:
-                            ECPoint pk1 = ECC.secp256r1.getCurve().decodePoint(o1);
-                            ECPoint pk2 = ECC.secp256r1.getCurve().decodePoint(o2);
-                            return ECC.compare(pk1, pk2);
-                        case EDDSA:
-                            //TODO
-                            return Helper.toHexString(o1).compareTo(Helper.toHexString(o1));
-                        default:
-                            return Helper.toHexString(o1).compareTo(Helper.toHexString(o1));
-                    }
-                }
-            });
-
-            for (byte[] publicKey : publicKeys) {
-                sb.push(publicKey);
-            }
-            sb.push(BigInteger.valueOf(publicKeys.length));
-            sb.add(ScriptOp.OP_CHECKMULTISIG);
-            return new Address(Digest.hash160(sb.toArray()));
-        }
+        return Address.toScriptHash(Program.ProgramFromMultiPubKey(m,publicKeys));
     }
+
 
     public static Address decodeBase58(String address) throws Exception {
         byte[] data = Base58.decode(address);
